refactor: signing module
extract the signing flow from main.py, and form a new package called
ncu_hsys.

the signing flow will now be execute from the package ncu_hsys.

after this change, if you want to do the signing flow test, you need
to call the function with `python -m "ncu_hsys.sign_bot"`.
